refactor(suite): fix AssetData type-only circular dep
